feat(frontend): add live run timeline
