+++ frontend/src/app/runs/[run_id]/page.tsx
@@ -1,0 +1,99 @@
+"use client";
+import { useEffect, useRef, useState } from "react";
+import RunTimeline, { TimelineStep } from "@/components/RunTimeline";
+import { http } from "@/lib/api";
+import { connectWS } from "@/lib/ws";
+
+interface Run {
+  run_id: string;
+  objective: string;
+  status: "running" | "done";
+  created_at: string;
+  completed_at?: string | null;
+  html?: string | null;
+  summary?: string | null;
+  steps: TimelineStep[];
+}
+
+export default function RunDetail({ params }: { params: { run_id: string } }) {
+  const [run, setRun] = useState<Run | null>(null);
+  const wsRef = useRef<WebSocket | null>(null);
+
+  useEffect(() => {
+    let cancelled = false;
+    async function load() {
+      const res = await http(`/runs/${params.run_id}`);
+      if (res.ok) {
+        const data = await res.json();
+        if (!cancelled) setRun(data);
+      }
+    }
+    load();
+
+    const ws = connectWS("/stream");
+    wsRef.current = ws;
+    ws.onopen = () => {
+      ws.send(JSON.stringify({ run_id: params.run_id }));
+    };
+    ws.onmessage = evt => {
+      const msg = JSON.parse(evt.data);
+      setRun(prev => {
+        if (!prev) return prev;
+        if (msg.status === "done") {
+          ws.close();
+          return {
+            ...prev,
+            status: "done",
+            html: msg.html ?? prev.html,
+            summary: msg.summary ?? prev.summary,
+            completed_at: msg.completed_at ?? prev.completed_at,
+          };
+        }
+        if (msg.node) {
+          const step: TimelineStep = {
+            order: msg.order ?? prev.steps.length + 1,
+            node: msg.node,
+            timestamp: msg.timestamp ?? new Date().toISOString(),
+            content: msg.content ?? "",
+          };
+          return { ...prev, steps: [...prev.steps, step] };
+        }
+        return prev;
+      });
+    };
+    ws.onerror = () => ws.close();
+    return () => {
+      cancelled = true;
+      ws.close();
+    };
+  }, [params.run_id]);
+
+  if (!run) return <div className="p-6">Loading...</div>;
+
+  return (
+    <div className="p-6 space-y-4">
+      <div className="space-y-1">
+        <h1 className="text-xl font-semibold">Run {run.run_id}</h1>
+        <p className="text-sm text-muted-foreground">Objective: {run.objective}</p>
+        <p className="text-sm text-muted-foreground">Status: {run.status}</p>
+        <p className="text-sm text-muted-foreground">
+          Started: {new Date(run.created_at).toLocaleString()}
+        </p>
+        {run.completed_at && (
+          <p className="text-sm text-muted-foreground">
+            Ended: {new Date(run.completed_at).toLocaleString()}
+          </p>
+        )}
+      </div>
+      <RunTimeline steps={run.steps} />
+      {run.status === "done" && (
+        <div className="space-y-4">
+          {run.summary && <p>{run.summary}</p>}
+          {run.html && (
+            <div className="prose" dangerouslySetInnerHTML={{ __html: run.html }} />
+          )}
+        </div>
+      )}
+    </div>
+  );
+}
