+++ frontend/src/components/RunTimeline.tsx
@@ -1,52 +1,57 @@
 "use client";
-import Link from "next/link";
+import { useState } from "react";
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog";
 
-interface Step {
-  step: string;
-  start: string;
-  end: string;
+export interface TimelineStep {
+  order: number;
+  node: string;
+  timestamp: string;
+  content: string;
 }
 
-interface Run {
-  run_id: string;
-  status: string;
-  steps?: Step[];
+interface TimelineProps {
+  steps: TimelineStep[];
 }
 
-const colors: Record<string, string> = {
-  plan: "bg-blue-400",
-  execute: "bg-yellow-400",
-  write: "bg-green-400",
-};
-
-export default function RunTimeline({ run }: { run: Run }) {
-  const steps = run.steps ?? [];
-  const durations = steps.map(
-    s => new Date(s.end).getTime() - new Date(s.start).getTime(),
-  );
-  const total = durations.reduce((a, b) => a + b, 0) || 1;
+export default function RunTimeline({ steps }: TimelineProps) {
+  const [selected, setSelected] = useState<TimelineStep | null>(null);
   return (
-    <div className="space-y-1">
-      <div className="flex h-2 w-full overflow-hidden rounded">
-        {steps.length === 0 ? (
-          <div
-            data-testid="timeline-loading"
-            className="h-full w-full animate-pulse rounded bg-gray-200"
-          />
-        ) : (
-          steps.map((s, i) => (
-            <div
-              key={s.step}
-              className={`${colors[s.step] || "bg-gray-400"} h-full`}
-              style={{ width: `${(durations[i] / total) * 100}%` }}
-              title={s.step}
-            />
-          ))
+    <div>
+      <ul className="divide-y rounded border">
+        {steps.map((s) => (
+          <li
+            key={s.order}
+            className="p-2 cursor-pointer hover:bg-muted/50"
+            onClick={() => setSelected(s)}
+          >
+            <div className="flex items-center gap-2">
+              <span className="font-mono text-xs w-6">{s.order}</span>
+              <span className="text-sm font-medium flex-1">{s.node}</span>
+              <span className="text-xs text-muted-foreground">
+                {new Date(s.timestamp).toLocaleTimeString()}
+              </span>
+            </div>
+            <p className="text-xs text-muted-foreground truncate">
+              {s.content}
+            </p>
+          </li>
+        ))}
+        {steps.length === 0 && (
+          <li className="p-2 text-sm text-muted-foreground">No steps yet</li>
         )}
-      </div>
-      <Link href={`/runs/${run.run_id}`} className="text-xs text-blue-500 hover:underline">
-        Détails du run
-      </Link>
+      </ul>
+      <Dialog open={!!selected} onOpenChange={(o) => !o && setSelected(null)}>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>
+              Step {selected?.order}: {selected?.node}
+            </DialogTitle>
+          </DialogHeader>
+          <div className="mt-2 whitespace-pre-wrap text-sm">
+            {selected?.content}
+          </div>
+        </DialogContent>
+      </Dialog>
     </div>
   );
 }
